--- layouts/MainLayout.module.scss
@@ -1,97 +1,0 @@
-@import '../styles/theme.scss';
-
-.header {
-  position: fixed;
-  width: 100%;
-  z-index: 9999;
-  .container {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-  .mainLogo {
-    cursor: pointer;
-    img {
-      width: 10rem;
-    }
-  }
-  nav {
-    display: flex;
-    gap: 2rem;
-    ul {
-      display: flex;
-      align-items: center;
-      gap: 3rem;
-      color: $white;
-      font-size: $font-size-lg;
-      li {
-        position: relative;
-        padding: 0.2rem;
-        a {
-          cursor: pointer;
-        }
-        &::before {
-          content: '';
-          position: absolute;
-          left: 0;
-          bottom: -0.2rem;
-          width: 0%;
-          height: 0.1rem;
-          background-color: $white;
-          transition: 0.5s;
-        }
-        &:hover {
-          &::before {
-            width: 100%;
-          }
-        }
-      }
-    }
-    .loginWrap {
-      display: flex;
-      gap: 1rem;
-      button {
-        font-size: $font-size-sm;
-        color: $white;
-        padding: 0;
-        outline: none;
-      }
-    }
-  }
-}
-
-@media screen and (min-width: $breakpoint-lg) {
-  .container {
-    width: 1060px;
-    margin: 2rem auto 3rem auto;
-  }
-}
-
-@media screen and (max-width: $breakpoint-lg) and (min-width: $breakpoint-sm) {
-  .container {
-    width: 100%;
-    margin: 2rem auto 3rem auto;
-    padding: 0 2rem;
-    nav {
-      display: none;
-    }
-  }
-}
-
-@media screen and (max-width: $breakpoint-sm) {
-  .header {
-    .mainLogo {
-      img {
-        width: 6rem;
-      }
-    }
-  }
-  .container {
-    width: 100%;
-    margin: 2rem auto 3rem auto;
-    padding: 0 2rem;
-    nav {
-      display: none;
-    }
-  }
-}
